Split out skill gui into separate files to reduce inclusion
