+++ fheroes2.vcxproj
@@ -301,6 +301,7 @@
     <ClCompile Include="src\fheroes2\gui\interface_icons.cpp" />
     <ClCompile Include="src\fheroes2\gui\interface_radar.cpp" />
     <ClCompile Include="src\fheroes2\gui\interface_status.cpp" />
+    <ClCompile Include="src\fheroes2\gui\skill_bar.cpp" />
     <ClCompile Include="src\fheroes2\gui\splitter.cpp" />
     <ClCompile Include="src\fheroes2\gui\statusbar.cpp" />
     <ClCompile Include="src\fheroes2\gui\text.cpp" />
@@ -455,6 +456,7 @@
     <ClInclude Include="src\fheroes2\gui\interface_list.h" />
     <ClInclude Include="src\fheroes2\gui\interface_radar.h" />
     <ClInclude Include="src\fheroes2\gui\interface_status.h" />
+    <ClInclude Include="src\fheroes2\gui\skill_bar.h" />
     <ClInclude Include="src\fheroes2\gui\splitter.h" />
     <ClInclude Include="src\fheroes2\gui\statusbar.h" />
     <ClInclude Include="src\fheroes2\gui\text.h" />
