+++ src/fheroes2/heroes/heroes_meeting.cpp
@@ -33,6 +33,7 @@
 #include "heroes.h"
 #include "heroes_indicator.h"
 #include "settings.h"
+#include "skill_bar.h"
 #include "text.h"
 
 void RedrawPrimarySkillInfo( const Point &, PrimarySkillsBar *, PrimarySkillsBar * );
