+++ src/fheroes2/kingdom/kingdom_overview.cpp
@@ -33,7 +33,7 @@
 #include "interface_list.h"
 #include "kingdom.h"
 #include "settings.h"
-#include "skill.h"
+#include "skill_bar.h"
 #include "text.h"
 #include "world.h"
